% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.ptsr}
\alias{print.ptsr}
\title{Print Method of class PTSR}
\usage{
\method{print}{ptsr}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{object of class \code{ptsr}.}

\item{digits}{minimal number of significant digits, see
\code{\link{print.default}}.}

\item{...}{further arguments to be passed to or from other methods.
They are ignored in this function}
}
\value{
Invisibly returns its argument, \code{x}.
}
\description{
Print method for objects of class \code{ptsr}.
}
\details{
Users are not encouraged to call these internal functions directly.
Internal functions for package PTSR.
}
